fix accuracy

+++ executor.py
@@ -334,7 +334,7 @@ class Executor(object):
         if self._accuracy is None:
             predictions = tf.to_int32(
                 graph['class_pred'] > self.config.default('accuracy_threshold', 0.5))
-            equality = tf.equal(predictions, graph['class_y'])
+            equality = tf.equal(predictions, tf.cast(graph['class_y'], tf.int32))
             self._accuracy = tf.reduce_mean(tf.cast(equality, tf.float32))
         return self._accuracy
